requirements.txt
Flask==3.1.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
eventlet==0.35.2
